+++ changelog.md
@@ -4,6 +4,9 @@
 * Check shulker boxes for items to place
 * Place blocks in the hot bar like a Quark Trowel
 
+## Upcoming
+* Mod can now run on a server
+
 ## 1.2.0
 * Added Angel Building. With this new Ability you can place blocks from your off-hand in midair
 * Fixed a bug where Construction wouldn't work if the durability of the tools was 1